Added www/p5-Catalyst-Plugin-AutoRestart version 0.92
Catalyst plugin to force the application to restart server processes
when they reach a configurable memory threshold. Memory checks are
performed every 'N' requests. This is intended as a band-aid to
deal with problems like memory leaks; it's here to buy you time to
find and solve the underlying issues.