ghostscript has hand-rolled shared lib handling in src/unix-dll.mak that does not do the right thing on Darwin (where shared libs are created with different linking flags and use dylib rather than so as an extension).
you need to manually reconfigure unix-dll.mak to make it work, so
do it in hacks.mk.

XXX: this still leaves the issue of libgs.so (et al.) being hardwired
in PLIST, instead of using libgs.dylib (et al.).   i'm not sure of
the best way to fix that.   at least it compiles now....